requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.1
google-auth-oauthlib


# Supabase
supabase==2.10.0

# HTTP client (pinned to avoid proxy TypeError)
httpx==0.27.2

# Twilio
twilio==8.12.0

# Email (Resend)
resend==2.0.0
Jinja2==3.1.3

# PDF processing
PyMuPDF==1.23.21
Pillow==10.2.0
reportlab==4.1.0
qrcode==7.4.2

# Security
PyJWT==2.8.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<0.24.0
pytest-cov>=4.0.0

# Development
python-dotenv==1.0.1

# Google Cloud Logging
google-cloud-logging==3.9.0

# Generative AI
google-generativeai==0.4.0
